@@ -1,5 +1,4 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { PersistGate } from "redux-persist/integration/react";
@@ -9,13 +8,14 @@ import App from './App';
 
 import { store, persistor } from './redux/store';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
     <Provider store={store}>
         <BrowserRouter>
             <PersistGate persistor={persistor}>
                 <App />
             </PersistGate>
         </BrowserRouter>
-    </Provider>,
-    document.getElementById('root')
+    </Provider>
 );
